refactored ItemAdapter to ItemList; added some comments
