+++ app/src/main/java/com/example/wongtonsoup/ItemList.java
@@ -1,74 +1,123 @@
 package com.example.wongtonsoup;
 
 import android.content.Context;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
 
-/**
- * Class for the list of item
- * @author yihui
- * @version 1.0
- * @since 10/28/2023
- */
 public class ItemList extends ArrayAdapter<Item> {
 
-    private final Context context;
-    private final ArrayList<Item> items;
-    public ItemList(Context context, ArrayList<Item> items) {
-        super(context, 0, items);
-        this.context = context;
-        this.items = items;
+    private Context mContext;
+    private List<Item> itemList; // The adapter's own list for display
+    private static ItemListListener listener; // Listener for adapter changes
+
+    // Interface for a class that wants to implement a listener on this class
+    public interface ItemListListener {
+        void onItemListChanged();
     }
 
-    @Override
-    public View getView(int position, @Nullable View convertView, @NonNull ViewGroup
-            parent){
+    // Constructor for the ItemAdapter
+    public ItemList(@NonNull Context context, @NonNull List<Item> objects) {
+        super(context, 0, new ArrayList<Item>()); // Initialize ArrayAdapter with an empty list
+        mContext = context;
+        itemList = new ArrayList<>(objects); // Create a separate list for the adapter
+        addAll(itemList); // Add all items to the ArrayAdapter for display
 
-        View view = convertView;
-        if(view == null){
-            view = LayoutInflater.from(context).inflate(R.layout.content, parent,false);
+    }
+
+    public static void setListener(ItemListListener customlistener) {
+        listener = customlistener;
+    }
+
+    // Returns the view for an item in the list
+    @NonNull
+    @Override
+    public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+        // Check if an existing view is being reused, otherwise inflate the view
+        if (convertView == null) {
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.content, parent, false);
         }
 
-        // Get the item at the current position
-        Item currentItem = items.get(position);
+        Item currentItem = getItem(position);
 
-        // Log position and other relevant information
-        Log.d("ItemAdapter", "getView - position: " + position);
+        TextView descriptionTextView = convertView.findViewById(R.id.content_description);
+        TextView dateTextView = convertView.findViewById(R.id.content_date);
+        TextView makeTextView = convertView.findViewById(R.id.content_make);
+        TextView priceTextView = convertView.findViewById(R.id.content_price);
 
-        // Set the image, content, and tags in the layout
-        assert convertView != null;
-//        ImageView imageView = view.findViewById(R.id.photo);
-//        imageView.setImageResource(currentItem.getImageResource());
+        descriptionTextView.setText(currentItem.getDescription());
+        dateTextView.setText(currentItem.getPurchaseDate().toString());
+        makeTextView.setText(currentItem.getMake());
+        priceTextView.setText(String.format(Locale.getDefault(), "%.2f", currentItem.getValue()));
 
-        TextView textViewName = view.findViewById(R.id.content_description);
-        textViewName.setText(currentItem.getDescription());
-
-        TextView textViewDate = view.findViewById(R.id.content_date);
-        textViewDate.setText(currentItem.getPurchaseDate());
-
-        TextView textViewMake = view.findViewById(R.id.content_make);
-        textViewMake.setText(currentItem.getMake());
-
-        TextView testViewPrice = view.findViewById(R.id.content_price);
-        testViewPrice.setText(currentItem.getValueAsString());
-
-//        TextView textViewTag1 = convertView.findViewById(R.id.textViewTag1);
-//        textViewTag1.setText(currentItem.getTag1());
-//
-//        TextView textViewTag2 = convertView.findViewById(R.id.textViewTag2);
-//        textViewTag2.setText(currentItem.getTag2());
-
-        return view;
+        // Return the completed view to render on screen
+        return convertView;
     }
 
+    /**
+     * Updates the list of items currently being displayed
+     * @param items new list to be displayed
+     */
+    public void updateData(List<Item> items) {
+        itemList.clear(); // Clear the adapter's own list
+        itemList.addAll(items); // Add new items to the adapter's list
+        clear(); // Clear the ArrayAdapter's internal list
+        addAll(itemList); // Add the adapter's list to the ArrayAdapter's internal list
+        notifyDataSetChanged(); // Notify the adapter that the data set has changed
+        if (listener != null){
+            listener.onItemListChanged();
+        }
+    }
+
+    public String getTotalDisplayed(){
+        double total = 0;
+        for (int i = 0 ; i < itemList.size() ; i++){
+            total += itemList.get(i).getValue();
+        }
+        String s = String.format("%.2f", total);
+        return s;
+    }
+
+    /**
+     * Searchs itemList by date
+     * @return sorted list.
+     */
+    public List<Item> sortByDate(){
+        itemList.sort(Item.byDate); // use the comparator in the Item class
+        return itemList;
+    }
+    /**
+     * Searchs itemList by description
+     * @return sorted list.
+     */
+    public List<Item> sortByDescription(){
+        itemList.sort(Item.byDescription); // use the comparator in the Item class
+        return itemList;
+    }
+    /**
+     * Searchs itemList by make
+     * @return sorted list.
+     */
+    public List<Item> sortByMake(){
+        itemList.sort(Item.byMake); // use the comparator in the Item class
+        return itemList;
+    }
+    /**
+     * Searchs itemList by value
+     * @return sorted list.
+     */
+    public List<Item> sortByValue(){
+        itemList.sort(Item.byValue); // use the comparator in the Item class
+        return itemList;
+    }
 
 }
+
